Initial commit
FFmpeg batch scripts for processing video and audio with x265 and Opus for smaller file sizes.
Re-encode - Opus is just for processing audio. It is currently set to use the opusenc application rather than FFmpeg's libopus. This means you must have opusenc.exe on your system and point this script to its location. (I use a folder within an ffmpeg folder.)
Re-encode - x265 + Opus is for processing video and audio. It uses Opus for the audio but the libopus version for convenience. Also the lone advantage I know for opusenc is not relevant here. The script is set to limit x265's bit rate based on the height of the input video.
Re-encode - x265 + Opus - multi is similar but for multiple audio tracks. This requires editing the file to match the number of audio tracks, their number of channels, and the titles you want for these tracks. It is currently configured based on my copy of The Matrix, after pulling just the English audio, including the four commentary tracks.