Audio Bitrate Control and Video Cropping
Have added the RATEopus function that can set a bitrate for libopus, rather than relying on the default values. You give it the bitrate you want for coupled channel pairs (stereo) and uncoupled channels (mono) and it does the math based on the number of channels. (5.1 is 6 channels: 2 pairs (front and rear stereo); two mono channels (center and subwoofer)). There is a maximum enforced of 512 Kbps.
Also added cropping of the video as I have run into some videos that needed this. Assuming the purpose is to remove letter boxing, the value you pass it is the height of one of the bars. It calculates the rest for you.