Audio Bitrate Control
Have added the RATEopusChan function that can set a bitrate for libopus, rather than relying on the default values. You give it the number of channels, the bitrate you want for coupled channel pairs (stereo) and uncoupled channels (mono) and it does the math based on the number of channels. (5.1 is 6 channels: 2 pairs (front and rear stereo); two mono channels (center and subwoofer)). There is a maximum enforced of 512 Kbps.
If either bitrate you provide is 0, however, no value is provided and libopus will use its defaults, which are the defaults for RATEopusChan too.